+++ OgreKit.framework/Versions/A/Headers/OgreTableView.h
@@ -3,8 +3,8 @@
  * Project: OgreKit
  *
  * Creation Date: Jun 13 2004
- * Author: Isao Sonobe <sonoisa (AT) muse (DOT) ocn (DOT) ne (DOT) jp>
- * Copyright: Copyright (c) 2003 Isao Sonobe, All rights reserved.
+ * Author: Isao Sonobe <sonoisa@gmail.com>
+ * Copyright: Copyright (c) 2003-2018 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -17,18 +17,18 @@
 
 @interface OgreTableView : NSTableView <OgreView>
 {
-    int     _ogreSelectedColumn;
-    int     _ogreSelectedRow;
-    NSRange _ogreSelectedRange;
+    NSInteger   _ogreSelectedColumn;
+    NSInteger   _ogreSelectedRow;
+    NSRange     _ogreSelectedRange;
 }
 
 - (NSObject <OgreTextFindComponent>*)ogreAdapter;
 
-- (int)ogreSelectedColumn;
-- (void)ogreSetSelectedColumn:(int)column;
+- (NSInteger)ogreSelectedColumn;
+- (void)ogreSetSelectedColumn:(NSInteger)column;
 
-- (int)ogreSelectedRow;
-- (void)ogreSetSelectedRow:(int)row;
+- (NSInteger)ogreSelectedRow;
+- (void)ogreSetSelectedRow:(NSInteger)row;
 
 - (NSRange)ogreSelectedRange;
 - (void)ogreSetSelectedRange:(NSRange)aRange;
