+++ OgreKit.framework/Versions/A/Headers/OgreTextFindBranch.h
@@ -3,8 +3,8 @@
  * Project: OgreKit
  *
  * Creation Date: Sep 26 2003
- * Author: Isao Sonobe <sonoisa (AT) muse (DOT) ocn (DOT) ne (DOT) jp>
- * Copyright: Copyright (c) 2003 Isao Sonobe, All rights reserved.
+ * Author: Isao Sonobe <sonoisa@gmail.com>
+ * Copyright: Copyright (c) 2003-2018 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -18,7 +18,7 @@
 @interface OgreTextFindBranch : NSObject <OgreTextFindComponent>
 {
     OgreTextFindBranch      *_parent;
-    int                     _index;
+    NSInteger               _index;
     BOOL                    _isParentRetained;
     BOOL                    _isTerminal;
     BOOL                    _isReversed;
