+++ OgreKit.framework/Versions/A/Headers/OgreTextFindComponent.h
@@ -3,8 +3,8 @@
  * Project: OgreKit
  *
  * Creation Date: Sep 26 2003
- * Author: Isao Sonobe <sonoisa (AT) muse (DOT) ocn (DOT) ne (DOT) jp>
- * Copyright: Copyright (c) 2003 Isao Sonobe, All rights reserved.
+ * Author: Isao Sonobe <sonoisa@gmail.com>
+ * Copyright: Copyright (c) 2003-2018 Isao Sonobe, All rights reserved.
  * License: OgreKit License
  *
  * Encoding: UTF8
@@ -37,15 +37,15 @@
 /* Getting and setting structural detail */
 - (BOOL)isLeaf;
 - (BOOL)isBranch;
-- (unsigned)numberOfChildrenInSelection:(BOOL)inSelection;
-- (unsigned)numberOfDescendantsInSelection:(BOOL)inSelection;
-- (id)childAtIndex:(unsigned)index inSelection:(BOOL)inSelection;
+- (NSUInteger)numberOfChildrenInSelection:(BOOL)inSelection;
+- (NSUInteger)numberOfDescendantsInSelection:(BOOL)inSelection;
+- (id)childAtIndex:(NSUInteger)index inSelection:(BOOL)inSelection;
 
 - (OgreTextFindBranch*)parent;
 - (void)setParent:(OgreTextFindBranch*)parent;
 - (void)setParentNoRetain:(OgreTextFindBranch*)parent;
-- (int)index;
-- (void)setIndex:(int)index;
+- (NSInteger)index;
+- (void)setIndex:(NSInteger)index;
 - (OgreTextFindLeaf*)selectedLeaf;
 
 - (BOOL)isTerminal;
